+++ src/info.rs
@@ -12,6 +12,7 @@ pub struct Info {
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Kind {
     Integer,
+    NonZeroInteger,
     Float,
     String,
     BoxStr,
@@ -24,6 +25,7 @@ impl Kind {
     pub fn is_clone(&self) -> bool {
         match self {
             Kind::Integer
+            | Kind::NonZeroInteger
             | Kind::Float
             | Kind::String
             | Kind::BoxStr
@@ -35,7 +37,7 @@ impl Kind {
 
     pub fn is_copy(&self) -> bool {
         match self {
-            Kind::Integer | Kind::Float | Kind::Char => true,
+            Kind::Integer | Kind::NonZeroInteger | Kind::Float | Kind::Char => true,
             Kind::String | Kind::BoxStr | Kind::StaticStr | Kind::Other => false,
         }
     }
@@ -43,6 +45,7 @@ impl Kind {
     pub fn is_debug(&self) -> bool {
         match self {
             Kind::Integer
+            | Kind::NonZeroInteger
             | Kind::Float
             | Kind::String
             | Kind::BoxStr
@@ -60,34 +63,54 @@ impl Kind {
             | Kind::BoxStr
             | Kind::StaticStr
             | Kind::Char => true,
-            Kind::Other => false,
+            Kind::NonZeroInteger | Kind::Other => false,
         }
     }
 
     pub fn is_deserialize(&self) -> bool {
         match self {
-            Kind::Integer | Kind::Float | Kind::String | Kind::BoxStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::Float
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::Char => true,
             Kind::Other | Kind::StaticStr => false,
         }
     }
 
     pub fn is_eq(&self) -> bool {
         match self {
-            Kind::Integer | Kind::String | Kind::BoxStr | Kind::StaticStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::StaticStr
+            | Kind::Char => true,
             Kind::Float | Kind::Other => false,
         }
     }
 
     pub fn is_hash(&self) -> bool {
         match self {
-            Kind::Integer | Kind::String | Kind::BoxStr | Kind::StaticStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::StaticStr
+            | Kind::Char => true,
             Kind::Float | Kind::Other => false,
         }
     }
 
     pub fn is_ord(&self) -> bool {
         match self {
-            Kind::Integer | Kind::String | Kind::BoxStr | Kind::StaticStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::StaticStr
+            | Kind::Char => true,
             Kind::Float | Kind::Other => false,
         }
     }
@@ -95,6 +118,7 @@ impl Kind {
     pub fn is_partial_eq(&self) -> bool {
         match self {
             Kind::Integer
+            | Kind::NonZeroInteger
             | Kind::Float
             | Kind::String
             | Kind::BoxStr
@@ -107,6 +131,7 @@ impl Kind {
     pub fn is_partial_ord(&self) -> bool {
         match self {
             Kind::Integer
+            | Kind::NonZeroInteger
             | Kind::Float
             | Kind::String
             | Kind::BoxStr
@@ -118,7 +143,12 @@ impl Kind {
 
     pub fn is_serialize(&self) -> bool {
         match self {
-            Kind::Integer | Kind::Float | Kind::String | Kind::BoxStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::Float
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::Char => true,
             Kind::Other | Kind::StaticStr => false,
         }
     }
@@ -126,6 +156,7 @@ impl Kind {
     pub fn is_display(&self) -> bool {
         match self {
             Kind::Integer
+            | Kind::NonZeroInteger
             | Kind::Float
             | Kind::String
             | Kind::BoxStr
@@ -137,7 +168,12 @@ impl Kind {
 
     pub fn is_from_str(&self) -> bool {
         match self {
-            Kind::Integer | Kind::Float | Kind::String | Kind::BoxStr | Kind::Char => true,
+            Kind::Integer
+            | Kind::NonZeroInteger
+            | Kind::Float
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::Char => true,
             Kind::Other | Kind::StaticStr => false,
         }
     }
@@ -145,14 +181,19 @@ impl Kind {
     pub fn is_number(&self) -> bool {
         match self {
             Kind::Integer | Kind::Float => true,
-            Kind::String | Kind::BoxStr | Kind::StaticStr | Kind::Char | Kind::Other => false,
+            Kind::NonZeroInteger
+            | Kind::String
+            | Kind::BoxStr
+            | Kind::StaticStr
+            | Kind::Char
+            | Kind::Other => false,
         }
     }
 
     pub fn is_string(&self) -> bool {
         match self {
             Kind::String | Kind::BoxStr | Kind::StaticStr => true,
-            Kind::Integer | Kind::Float | Kind::Char | Kind::Other => false,
+            Kind::Integer | Kind::NonZeroInteger | Kind::Float | Kind::Char | Kind::Other => false,
         }
     }
 }